feat: 비즈니스 로직을 LottoService 클래스로 이동
- 리뷰어 피드백 반영 (https://github.com/woowacourse/javascript-lotto/pull/370#discussion_r1965681147)
